fix: address round 6 review comments (5 items)
Bug fixes:
- Monitor: bind make_interval params as i64 instead of f64 (Postgres type safety)
- Notifications: mark_read returns NotFound when no rows updated

Documentation:
- Test harness docs: replaced 'testcontainers' with 'docker compose' (matches reality)

+++ backend/crates/monitor/src/service.rs
@@ -135,7 +135,7 @@ impl IdleMonitor {
     /// Return all running agents that have been idle longer than
     /// `idle_timeout`.
     async fn find_idle_agents(&self) -> anyhow::Result<Vec<Agent>> {
-        let timeout_minutes = self.idle_timeout.as_secs() as i64 / 60;
+        let timeout_minutes = (self.idle_timeout.as_secs() / 60) as i64;
 
         let agents: Vec<Agent> = sqlx::query_as(
             "SELECT * FROM agents \
@@ -143,7 +143,7 @@ impl IdleMonitor {
              AND (last_active IS NULL \
                   OR last_active < NOW() - make_interval(mins => $1))",
         )
-        .bind(timeout_minutes as f64)
+        .bind(timeout_minutes)
         .fetch_all(&self.db)
         .await
         .context("Failed to query idle agents")?;
@@ -188,7 +188,7 @@ impl IdleMonitor {
              AND next_run_at < NOW() + make_interval(mins => $2)",
         )
         .bind(agent_id)
-        .bind(UPCOMING_JOB_WINDOW_MINUTES as f64)
+        .bind(UPCOMING_JOB_WINDOW_MINUTES)
         .fetch_one(&self.db)
         .await
         .context("Failed to check upcoming scheduled jobs")?;
